test: split visual blank property regression
